prefabs spawn in right position
movement key presses recognized by the debug.log console. and prefabs not spawning at z=100 now

+++ EscapeRoom/Assets/Scripts/UIManager.cs
@@ -83,7 +83,7 @@ public class UIManager : MonoBehaviour
     // a unsername is more important and should require reliable messages. 
     public void SendName()
     {
-        Message message = Message.Create(MessageSendMode.reliable, (ushort)EscapeRoomClientToServerId.name);
+        Message message = Message.Create(MessageSendMode.reliable, EscapeRoomClientToServerId.name);
         message.AddString(usernameField.text);
         NetworkManager.Singleton.EscaperoomClient.Send(message);
     }
